requirements.txt
boto3==1.9.115
dj_database_url == 0.5.0
Django==2.1.5
django-allauth==0.40.0
django-daterange-filter==1.3.0
django-filter==2.2.0
django-forms-bootstrap==3.1.0
django-materializecss-form==1.1.12
django-registration==3.0.1
django-rest-auth==0.9.5
django-storages==1.7.1
django-webpack-loader==0.6.0
gunicorn==19.9.0
DjangoRestFramework==3.9.1
numpy==1.18.1
Pillow==7.0.0
psycopg2-binary==2.8.4